Fix alignment bug causing crashes in the 512-bit arithmetic code.
While testing the avx2 optimized code on HardenedBSD, we were seeing spurious crashes
when running the test suite. These crashes manifested when exercising the CSIDH-512 module
from Python, but not in the CSIDH-1024 or CSIDH-2048 code.

Attaching a debugger revealed the culprit:
A segfault in this code:
```
->  0x2eca3896278 <+107>: movdqa (%rax), %xmm0     // pseudocode: uint128_t xmm0 = *rax;
    0x2eca389627c <+111>: movdqa 0x10(%rax), %xmm1
    0x2eca3896281 <+116>: movaps 0x20(%rax), %xmm2
    0x2eca3896285 <+120>: movaps 0x30(%rax), %xmm3
```

Which corresponds to the C code in validate.c:validate():
```
  fp tmp2 = fp_2;
```

This was somewhat curious as `fp_2` is a global symbol and the memory is readily available:
```
(lldb) register read
General Purpose Registers:
       rax = 0x000002eca387cfe8  highctidh_512.cpython-311.so`highctidh_512_fp_2
```

The problem turned out to be that for the smaller field sizes, clang on HardenedBSD would
emit `movdqa` instructions, which expects 16-byte alignment.
But `highctidh_512_fp_2` above ends in `cfe8` (not cfe0 or cff0), so it is not aligned,
and hence the crash. The larger field sizes were unaffected because clang emitted `rep movsb`
for these; as did GCC for all field sizes.

The problem turned out to be `__attribute__((aligned(32)))` annotations added to fp.h
 and uintbig.h as part of the commit that introduced the fiat-crypto code.

The fix chosen here is to remove those to revert to the default 8-byte alignment imposed by
the uint64_t array members of the affected structs.