v1.0.2023120600
This version replace symlinks with pristine file copy from `../` in the
highctidh git repository.

This is needed because of a limitation with golang.  The `go get`
command has a relatively undocumented limitation in that a git repo with
symlinks is only usable if it manually has been cloned with git, or if
there is an override to use ssh+git. Anonymous ssh+git is rare and so a
developer or user would require extra steps, something we want to avoid
at all costs. This is an annoying limitation and making a file copy is
an easy solution to ensure no extra steps are required for users of the
golang module.

To update the source code in the modules run:

    `make -f Makefile.packages update-golang-modules`

After syncing a selected set of files from upstream do not forget to
check in the changes and make a new tagged release.