Initial experimental Windows porting work
This branch adds a small set of changes that allow the library to be
functional on Windows when built with `clang`. This changeset also
enables a CI configuration that compiles and runs some of the C library
tests on Windows Server 2019 and Windows Server 2022.

The Python and Golang modules may or may not be functional. Do not use
this for anything important except as a basis for further porting work.
A serious security review is needed before this should be considered
worth using in any context -- you have been warned!