Experimental build target for WebAssembly (wasm)
This changeset adds experimental WebAssembly support which is completely
untested and it should not be used. This is assumed to be insecure and
broken until reviewed, audited, and extensively tested.

The top-level `Makefile` now has a `wasm` target to build
`dist/highctidh.wasm`. The `autogen` script has been updated to reflect
the changes to `src/Makefile`.

This changeset also adds a CI configuration `ubuntu-webassembly.yml` for
building the library with the `emcc` compiler provided by the
`emscripten` package on Ubuntu GNU/Linux. It calls a new script
`misc/install-ubuntu-webassembly-deps.sh` to install the required
packages in the CI.

The `README.md` has been updated to reflect the new target.

Do not use this for anything - you have been warned!